refactored bunch of functions made graph points line up with axes correctly added graph title
Note: When running from a file:// scheme,
Chrome does not treat files in the same directory as same-origin,
so loading the movies.csv file will cause an error.
Firefox and Edge, however, implement the same-origin policy
for the file:// scheme, and so it works on Firefox and Edge.
I'm trying to figure out how to make it work on Chrome.